feesix: rise of the runtime

+++ src/main/kotlin/engine/physics/PhysicsSystem.kt
@@ -7,5 +7,5 @@ interface PhysicsSystem {
 
     fun unregisterBody(entityId: String)
 
-    fun step(scene: Scene, dt: Float)
+    fun step(scene: Scene, dt: Double)
 }
